raw chart, refactor
